requirements.txt
numpy==1.23.5
tensorflow==2.12.0
tensorflow-datasets==4.9.2
keras==2.12.0
Pillow==9.4.0
